backend/requirements.txt
# NyayaAI Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Environment & Configuration
python-dotenv==1.0.1
pydantic-settings==2.5.2

# LangChain & RAG
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2

# Embeddings
sentence-transformers==3.3.1

# Vector Store
faiss-cpu==1.9.0.post1

# PDF Processing
pypdf==5.1.0

# HTTP Client (for OpenRouter calls)
httpx==0.27.2

# Rate Limiting
slowapi==0.1.9

# Logging
loguru==0.7.3
